Move tasks to appropriate flow modules
HEDWG2-1211
