+++ em_workflows/brt/config.py
@@ -1,4 +1,5 @@
 import os
+from collections import namedtuple
 
 from em_workflows.config import Config
 
@@ -6,3 +7,6 @@ from em_workflows.config import Config
 class BRTConfig(Config):
     binvol = os.environ.get("BINVOL_LOC", "/opt/rml/imod/bin/binvol")
     clip_loc = os.environ.get("CLIP_LOC", "/opt/rml/imod/bin/clip")
+
+
+BrtOutput = namedtuple("BrtOutput", ["ali_file", "rec_file"])
